requirements.txt
# Azure SDK
azure-identity>=1.15.0
azure-mgmt-datafactory>=4.0.0
azure-mgmt-monitor>=6.0.0
azure-mgmt-resource>=23.0.0
google-generativeai>=0.4.0
jinja2>=3.1.0
python-dotenv>=1.0.0
rich>=13.0.0
click>=8.1.0
flask>=3.0.0
gunicorn>=21.2.0
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Configuration
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
